uvtd: ctx: link control/legacy nodes with VTs
We now create a VT for each legacy node and assign all clients to it. For
control nodes, we create a different VT for each open-file/client.

Note that we do not recreate VTs or CDEVs directly on HUP instead of
during the next open() or reconfiguration. This avoids trapping into the
same error again and gives control to the user to recreate the nodes at
the right time.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>