wlt: theme: prevent move/resize requests when maximized/fullscreen
We should let the user move or resize the window while it is fullscreen or
maximized. This depends on the compositor to stop pending move/resize
requests when maximizing or setting a window fullscreen.

We can implement some "snap away from edges" behavior that allows to move
or resize while maximized. This will "unmaximize" the window if you move
it more than a given threshold. However, that's not needed now so lets do
that later when the wl_shell system is fully figured out.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>