font: pango: fix vertical alignment of non-ASCII glyphs
pango_ft2_render_layout_line() requires the baseline offset as argument,
not the vertical extent of the glyphs. This is important as we currently
align glyphs to the bottom edge instead of the baseline.

Fix this by passing the cached baseline offset so all glyphs are correctly
aligned to the baseline. We then clip according to the cell-extents as
usual.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>