shl: log: check for trailing newlines
If the format string contains a trailing newline, we now skip writing our
own. We also skip the fn+file information as it would be written on the
next line.

Trailing newlines occur only when forwarding messages from other
libraries. In this case we don't need the fn+file information, anyway.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>