wlt: toolkit: dont depend on implicit buffer releases
Wayland compositors are not required to send buffer release notifications
on an implicit buffer-attach. That is, an attach->commit->commit series
does not cause a buffer-release notification after the second commit.

Hence, send a buffer-attach on every commit that requires a new
buffer-release notification.

For the related wayland-protocol changes, see:
  http://cgit.freedesktop.org/wayland/wayland/commit/?id=7165bf21917ed1c4f5ec823a1e1642a9e83aa984

This fixes a bug where wlterm only redraws the window-content when
resizing (due to explicit buffer-attach calls during redraw).

Reported by: MoD
Signed-off-by: David Herrmann <dh.herrmann@gmail.com>