genunifont: fix MAX_DATA_SIZE to be in 8bit range
We only use uint8_t as length-counter so MAX_DATA_SIZE must be <256 to
make our length-checks work. This fixes a bug where we read invalid data
if someone uses modified genunifont data files. As this just produces
garbled glyphs if garbled data is given, this is actually not critical at
all.

Reported-by: Edward O'Callaghan <eocallaghan@alterapraxis.com>
Signed-off-by: David Herrmann <dh.herrmann@gmail.com>